+++ packages/data-loader/src/controllers/export.ts
@@ -10,9 +10,11 @@ import { printAsCsv } from "../printers/printAsCsv";
 export type Options = {
   app: AppID;
   attachmentDir?: string;
-  format?: "json" | "csv";
+  format?: ExportFileFormat;
   query?: string;
 };
+
+export type ExportFileFormat = "json" | "csv";
 
 type FileInfo = {
   name: string;