+++ packages/data-loader/src/controllers/import.ts
@@ -4,8 +4,8 @@ import {
 } from "@kintone/rest-api-client";
 import { AppID } from "@kintone/rest-api-client/lib/client/types";
 import { buildRestAPIClient, RestAPIClientOptions } from "../api";
-import { promises as fs } from "fs";
 import path from "path";
+import fs from "fs";
 import { parseJson } from "../parsers/parseJson";
 import { parseCsv } from "../parsers/parseCsv";
 import { ParsedRecord } from "../types";
@@ -40,13 +40,23 @@ export const buildImporter = ({
     return path.extname(filepath).split(".").pop() || "";
   };
 
+  async function readStream(stream: fs.ReadStream, encoding = "utf8") {
+    stream.setEncoding(encoding);
+    let content = "";
+    for await (const chunk of stream) {
+      content += chunk;
+    }
+    return content;
+  }
+
   async function importRecords(options: Options) {
     const { app, filePath } = options;
-    const buf = await fs.readFile(filePath);
+    const stream = fs.createReadStream(filePath);
+    const content = await readStream(stream);
     const type = extractFileType(filePath);
     const records = await parseSource({
       type,
-      source: buf.toString(),
+      source: content,
       options,
     });
     await addAllRecordsChunk(app, records);