+++ packages/customize-uploader/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.14](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.13...@kintone/customize-uploader@3.1.14) (2021-04-27)
+
+**Note:** Version bump only for package @kintone/customize-uploader
+
+
+
+
+
+## [3.1.13](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.12...@kintone/customize-uploader@3.1.13) (2021-04-20)
+
+**Note:** Version bump only for package @kintone/customize-uploader
+
+
+
+
+
+## [3.1.12](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.11...@kintone/customize-uploader@3.1.12) (2021-04-13)
+
+**Note:** Version bump only for package @kintone/customize-uploader
+
+
+
+
+
+## [3.1.11](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.10...@kintone/customize-uploader@3.1.11) (2021-04-06)
+
+
+### Bug Fixes
+
+* **customize-uploader:** pass guestSpaceId ([#799](https://github.com/kintone/js-sdk/issues/799)) ([9694d5f](https://github.com/kintone/js-sdk/commit/9694d5f08db86c14d2b31834b8dda5f815a1ce63))
+
+
+
+
+
+## [3.1.10](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.9...@kintone/customize-uploader@3.1.10) (2021-03-31)
+
+**Note:** Version bump only for package @kintone/customize-uploader
+
+
+
+
+
 ## [3.1.9](https://github.com/kintone/js-sdk/compare/@kintone/customize-uploader@3.1.8...@kintone/customize-uploader@3.1.9) (2021-03-23)
 
 **Note:** Version bump only for package @kintone/customize-uploader