+++ packages/webpack-plugin-kintone-plugin/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.22](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.21...@kintone/webpack-plugin-kintone-plugin@4.0.22) (2021-04-27)
+
+**Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin
+
+
+
+
+
+## [4.0.21](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.20...@kintone/webpack-plugin-kintone-plugin@4.0.21) (2021-04-20)
+
+**Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin
+
+
+
+
+
+## [4.0.20](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.19...@kintone/webpack-plugin-kintone-plugin@4.0.20) (2021-04-13)
+
+**Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin
+
+
+
+
+
+## [4.0.19](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.18...@kintone/webpack-plugin-kintone-plugin@4.0.19) (2021-04-06)
+
+**Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin
+
+
+
+
+
+## [4.0.18](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.17...@kintone/webpack-plugin-kintone-plugin@4.0.18) (2021-03-31)
+
+**Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin
+
+
+
+
+
 ## [4.0.17](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@4.0.16...@kintone/webpack-plugin-kintone-plugin@4.0.17) (2021-03-23)
 
 **Note:** Version bump only for package @kintone/webpack-plugin-kintone-plugin